% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtrick.R
\name{digest}
\alias{digest}
\alias{hmac}
\title{Raw hash functions}
\usage{
digest(x, m, algo)

hmac(x, m, algo)
}
\arguments{
\item{x}{character vector to hash}

\item{m}{modulo - an integer}

\item{algo}{\code{'md5'} and others provided by OpenSSL}

\item{key}{a raw vector. If provided, generates a different mapping}
}
\description{
These are C functions for use when sensitive to performance.
}

