% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtrick.R
\name{hashtrick}
\alias{hash_col}
\alias{hash_col.default}
\alias{hash_col.factor}
\alias{hashtrick}
\title{Hash Trick Functions for Model Building}
\usage{
hashtrick(..., m = 256L, algo = "md5", key)

hash_col(x, m, algo, key)

\method{hash_col}{default}(x, m, algo, key)

\method{hash_col}{factor}(x, m, algo, key)
}
\arguments{
\item{...}{a list of variables}

\item{m}{modulo - an integer}

\item{algo}{\code{'md5'} and others provided by OpenSSL}

\item{key}{a raw vector. If provided, the hash function generates a different mapping}

\item{x}{character vector to hash}
}
\description{
\code{hashtrick} builds a model matrix using the hash trick, \code{hash_col} hashes a single column.
}

