% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtrick.R
\name{hmac}
\alias{hmac}
\title{Raw hash functions}
\usage{
hmac(x, m, algo, key)
}
\arguments{
\item{x}{character vector to hash}

\item{m}{modulo - an integer}

\item{algo}{\code{'md5'} and others provided by OpenSSL}

\item{key}{a list of raw vectors. Each specifies a different mapping}
}
\description{
These are C functions for use when sensitive to performance.
}
