% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtrick.R
\name{hashtrick}
\alias{hashtrick}
\alias{hash_col}
\alias{hash_col.character}
\alias{hash_col.numeric}
\alias{hash_col.factor}
\title{Hash Trick Functions for Model Building}
\usage{
hashtrick(..., m = 256L, algo = "md5", key = DEFAULT_KEY)

hash_col(x, m, algo, key)

\method{hash_col}{character}(x, m, algo, key)

\method{hash_col}{numeric}(x, m, algo, key)

\method{hash_col}{factor}(x, m, algo, key)
}
\arguments{
\item{...}{a list of variables}

\item{m}{modulo - number of columns for result}

\item{algo}{\code{'md5'} and others provided by OpenSSL}

\item{key}{a list of raws. If provided, the hash function generates a different mapping}

\item{x}{character vector to hash}
}
\value{
for \code{hashtrick}, a n-by-m matrix of the dummy coded features. For \code{hash_col},
  a matrix containing hashes, one column per key provided.
}
\description{
\code{hashtrick} builds a matrix using the hash trick, \code{hash_col} hashes a single column.
}
